Refactor <GradingHometask /> component
